Replace Express server with Vercel functions
