+++ api/feedback.js
@@ -1,0 +1,44 @@
+import nodemailer from 'nodemailer';
+
+export default async function handler(req, res) {
+  if (req.method !== 'POST') {
+    res.setHeader('Allow', 'POST');
+    return res.status(405).end('Method Not Allowed');
+  }
+
+  try {
+    const { name, email, message } = req.body || {};
+
+    if (!name || !email || !message) {
+      return res.status(400).json({ error: 'Name, email, and message are required' });
+    }
+
+    if (!email.includes('@')) {
+      return res.status(400).json({ error: 'Please provide a valid email address' });
+    }
+
+    const transporter = nodemailer.createTransport({
+      host: process.env.SMTP_HOST,
+      port: Number(process.env.SMTP_PORT || 587),
+      secure: false,
+      auth: {
+        user: process.env.SMTP_USER,
+        pass: process.env.SMTP_PASS,
+      },
+    });
+
+    const textContent = `Name: ${name}\nEmail: ${email}\nMessage: ${message}`;
+
+    await transporter.sendMail({
+      from: process.env.FROM_EMAIL,
+      to: process.env.TO_EMAIL,
+      subject: `Vibe Pick: Feedback ${email}`,
+      text: textContent,
+    });
+
+    return res.status(200).json({ message: 'Feedback sent successfully' });
+  } catch (error) {
+    console.error('Error sending feedback:', error);
+    return res.status(500).json({ error: 'Failed to send feedback. Please try again later.' });
+  }
+}
