--- server/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "just-one-place-backend",
-  "version": "1.0.0",
-  "description": "Backend server for Just One Place app",
-  "main": "server.js",
-  "type": "module",
-  "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js"
-  },
-  "dependencies": {
-    "cors": "^2.8.5",
-    "dotenv": "^16.3.1",
-    "express": "^4.18.2",
-    "express-rate-limit": "^7.1.5",
-    "stripe": "^18.4.0"
-  },
-  "devDependencies": {
-    "nodemon": "^3.0.2"
-  }
-}
