--- server/test-email-direct.js
@@ -1,57 +1,0 @@
-import dotenv from 'dotenv';
-import path from 'path';
-import { fileURLToPath } from 'url';
-import nodemailer from 'nodemailer';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-
-// Load .env from the current directory
-dotenv.config();
-
-const testEmail = async () => {
-  try {
-    console.log('=== Direct Email Test ===');
-    console.log('SMTP_HOST:', process.env.SMTP_HOST);
-    console.log('SMTP_PORT:', process.env.SMTP_PORT);
-    console.log('SMTP_USER:', process.env.SMTP_USER);
-    console.log('SMTP_PASS:', process.env.SMTP_PASS ? '***' + process.env.SMTP_PASS.slice(-4) : 'NOT SET');
-    console.log('');
-    
-    console.log('Creating transporter...');
-    const transporter = nodemailer.createTransport({
-      host: process.env.SMTP_HOST,
-      port: process.env.SMTP_PORT,
-      auth: {
-        user: process.env.SMTP_USER,
-        pass: process.env.SMTP_PASS
-      }
-    });
-
-    console.log('Transporter created successfully');
-    
-    // Verify connection
-    console.log('Verifying connection...');
-    await transporter.verify();
-    console.log('✅ Connection verified successfully!');
-    
-    // Test sending email
-    console.log('Testing email send...');
-    const info = await transporter.sendMail({
-      from: process.env.FROM_EMAIL,
-      to: process.env.TO_EMAIL,
-      subject: 'Test Email - Just One Place',
-      text: 'This is a test email to verify the configuration is working.',
-      html: '<h1>Test Email</h1><p>This is a test email to verify the configuration is working.</p>'
-    });
-    
-    console.log('✅ Email sent successfully!');
-    console.log('Message ID:', info.messageId);
-    
-  } catch (error) {
-    console.error('❌ Error:', error.message);
-    console.error('Full error:', error);
-  }
-};
-
-testEmail();
